* commit '80e26fd3957e1c0f2ecb33cf744a790e4b6230e8':
  Optimze TimesharingplanChart UI show; Optimze stock quotation detail UI show;
  Optimze color resoure;

+++ src/com/dkhs/portfolio/bean/FSDataBean.java
@@ -84,18 +84,18 @@ public class FSDataBean {
      * */
 
     public class TimeStock {
-        private float volume;
+        private int volume;
         private float avgcurrent;
         private float avgline;
         private float current;
         private float percentage;
         private String time;
 
-        public float getVolume() {
+        public int getVolume() {
             return volume;
         }
 
-        public void setVolume(float volume) {
+        public void setVolume(int volume) {
             this.volume = volume;
         }
